refactor: move dn to vectors
